requirements.txt
Flask==3.1.1
Werkzeug==3.1.3
gunicorn==21.2.0
flask-cors==4.0.0
flask-limiter==3.5.0
pycryptodome==3.18.0
python-dotenv==1.0.0
requests==2.31.0